restructure; subscribe for remaining collections
