+++ bigbluebutton-html5/imports/api/chat/chatCollection.js
@@ -1,0 +1,1 @@
+export const Chat = new Mongo.Collection('chat');
