+++ bigbluebutton-html5/imports/api/chat/server/publications.js
@@ -1,0 +1,25 @@
+import { Chat } from '../chatCollection';
+import { isAllowedTo } from '/imports/startup/server/userPermissions';
+import { logger } from '/imports/startup/server/logger';
+
+Meteor.publish('chat', function (meetingId, userid, authToken) {
+  if (isAllowedTo('subscribeChat', meetingId, userid, authToken)) {
+    logger.info(`publishing chat for ${meetingId} ${userid} ${authToken}`);
+    return Chat.find({
+      $or: [
+        {
+          'message.chat_type': 'PUBLIC_CHAT',
+          meetingId: meetingId,
+        }, {
+          'message.from_userid': userid,
+          meetingId: meetingId,
+        }, {
+          'message.to_userid': userid,
+          meetingId: meetingId,
+        },
+      ],
+    });
+  } else {
+    this.error(new Meteor.Error(402, "The user was not authorized to subscribe for 'chats'"));
+  }
+});
