+++ bigbluebutton-html5/imports/api/cursor/cursorCollection.js
@@ -1,0 +1,1 @@
+export const Cursor = new Mongo.Collection('cursor');
