+++ bigbluebutton-html5/imports/api/cursor/server/publications.js
@@ -1,0 +1,9 @@
+import { Cursor } from '../cursorCollection';
+import { logger } from '/imports/startup/server/logger';
+
+Meteor.publish('cursor', function (meetingId) {
+  logger.info(`publishing cursor for ${meetingId}`);
+  return Cursor.find({
+    meetingId: meetingId,
+  });
+});
