+++ bigbluebutton-html5/imports/api/index.js
@@ -1,0 +1,18 @@
+import { Users } from './users/usersCollection';
+import { Meetings } from './meetings/meetingsCollection';
+import { Slides } from './slides/slidesCollection';
+import { Presentations } from './presentations/presentationsCollection';
+import { Shapes } from './shapes/shapesCollection';
+import { Chat } from './chat/chatCollection';
+import { Cursor } from './cursor/cursorCollection';
+
+
+export const Collections = {
+  Users: Users,
+  Meetings: Meetings,
+  Slides: Slides,
+  Presentations: Presentations,
+  Shapes: Shapes,
+  Chat: Chat,
+  Cursor: Cursor,
+};
