+++ bigbluebutton-html5/imports/api/meetings/meetingsCollection.js
@@ -1,0 +1,1 @@
+export const Meetings = new Mongo.Collection('meetings');
