+++ bigbluebutton-html5/imports/api/meetings/server/publications.js
@@ -1,0 +1,9 @@
+import { Meetings } from '../meetingsCollection';
+import { logger } from '/imports/startup/server/logger';
+
+Meteor.publish('meetings', function (meetingId) {
+  logger.info(`publishing meetings for ${meetingId}`);
+  return Meetings.find({
+    meetingId: meetingId,
+  });
+});
