+++ bigbluebutton-html5/imports/api/presentations/presentationsCollection.js
@@ -1,0 +1,1 @@
+export const Presentations = new Mongo.Collection('presentations');
