+++ bigbluebutton-html5/imports/api/shapes/server/publications.js
@@ -1,0 +1,7 @@
+import { Shapes } from '/imports/api/shapes/shapesCollection';
+
+Meteor.publish('shapes', function (meetingId) {
+  return Shapes.find({
+    meetingId: meetingId,
+  });
+});
