+++ bigbluebutton-html5/imports/api/shapes/shapesCollection.js
@@ -1,0 +1,1 @@
+export const Shapes = new Mongo.Collection('shapes');
