+++ bigbluebutton-html5/imports/api/slides/slidesCollection.js
@@ -1,0 +1,1 @@
+export const Slides = new Mongo.Collection('slides');
