+++ bigbluebutton-html5/imports/api/users/usersCollection.js
@@ -1,0 +1,1 @@
+export const Users = new Mongo.Collection('users');
