+++ bigbluebutton-html5/imports/startup/server/userPermissions.js
@@ -1,4 +1,4 @@
-import { Users, Meetings } from '/imports/startup/collections';
+import { Collections } from '/imports/api/index';
 import { logger } from '/imports/startup/server/logger';
 
 let moderator, presenter, viewer;
@@ -71,8 +71,8 @@ viewer = function (meetingId, userId) {
 
     // muting
     muteSelf: true,
-    unmuteSelf: !((meeting = Meetings.findOne({ meetingId: meetingId })) != null && meeting.roomLockSettings.disableMic) ||
-    !((user = Users.findOne({ meetingId: meetingId, userId: userId })) != null && user.user.locked),
+    unmuteSelf: !((meeting = Collections.Meetings.findOne({ meetingId: meetingId })) != null && meeting.roomLockSettings.disableMic) ||
+    !((user = Collections.Users.findOne({ meetingId: meetingId, userId: userId })) != null && user.user.locked),
 
     logoutSelf: true,
 
@@ -81,11 +81,11 @@ viewer = function (meetingId, userId) {
     subscribeChat: true,
 
     //chat
-    chatPublic: !((meeting = Meetings.findOne({ meetingId: meetingId })) != null && meeting.roomLockSettings.disablePublicChat) ||
+    chatPublic: !((meeting = Collections.Meetings.findOne({ meetingId: meetingId })) != null && meeting.roomLockSettings.disablePublicChat) ||
     !((user = Users.findOne({ meetingId: meetingId, userId: userId })) != null && user.user.locked) ||
     (user != null && user.user.presenter),
 
-    chatPrivate: !((meeting = Meetings.findOne({ meetingId: meetingId })) != null && meeting.roomLockSettings.disablePrivateChat) ||
+    chatPrivate: !((meeting = Collections.Meetings.findOne({ meetingId: meetingId })) != null && meeting.roomLockSettings.disablePrivateChat) ||
     !((user = Users.findOne({ meetingId: meetingId, userId: userId })) != null && user.user.locked) ||
     (user != null && user.user.presenter),
 
@@ -103,7 +103,7 @@ viewer = function (meetingId, userId) {
 // actions and the default value - see 'viewer' and 'moderator' in the beginning of the file
 export function isAllowedTo(action, meetingId, userId, authToken) {
   let user, validated;
-  user = Users.findOne({
+  user = Collections.Users.findOne({
     meetingId: meetingId,
     userId: userId,
   });
@@ -113,7 +113,7 @@ export function isAllowedTo(action, meetingId, userId, authToken) {
   return true; // TODO REMOVE THIS
 
   logger.info(`in isAllowedTo: action-${action}, userId=${userId}, authToken=${authToken} validated:${validated}`);
-  user = Users.findOne({
+  user = Collections.Users.findOne({
     meetingId: meetingId,
     userId: userId,
   });
